+++ packages/button/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @launchpad-ui/button
 
+## 0.12.15
+
+### Patch Changes
+
+- Updated dependencies [[`b255f1d`](https://github.com/launchdarkly/launchpad-ui/commit/b255f1dddf14c437585383cea5114ca93b949966)]:
+  - @launchpad-ui/icons@0.18.2
+
 ## 0.12.14
 
 ### Patch Changes