+++ packages/collapsible/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @launchpad-ui/collapsible
 
+## 0.1.59
+
+### Patch Changes
+
+- Updated dependencies [[`b255f1d`](https://github.com/launchdarkly/launchpad-ui/commit/b255f1dddf14c437585383cea5114ca93b949966)]:
+  - @launchpad-ui/icons@0.18.2
+  - @launchpad-ui/button@0.12.15
+
 ## 0.1.58
 
 ### Patch Changes